Refactor - Item Page의 data를 다루는 컴포넌트 별도로 분리
이유: Item 컴포넌트 폴터 내에 정의된 Inner 컴포넌트를 Page에서 호출하게 되면 폴더 거리가 멀어 사용성이 떨어집니다.
따라서 data를 처리하는 Item 컴포넌트를 별도로 분리하고 Inner 컴포넌트는 해당 컴포넌트에서 호출하도록 하여
호출 거리를 가깝게 수정하였습니다.
