+++ components/item/OptionViewer/OptionViewer.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement, useState } from 'react';
 import { Option } from 'apis/models/Items.type';
 import { commaNumber, dateFormatting } from 'utils';
-import { OptionSelector, SelectedOptionViewer } from '..';
+import { OptionSelector, SelectedOptionViewer } from '../innerPaths';
 import * as S from './OptionsViewer.styled';
 
 const OptionViewer = ({
