+++ components/item/SelectedOptionViewer/SelectedOptionViewer.styled.ts
