+++ components/item/TextsSection/TextsSection.type.ts
