+++ components/item/index.type.tsx
@@ -1,0 +1,5 @@
+import { ConItem } from 'apis/models/Items.type';
+
+export default interface ItemProps {
+  data: ConItem;
+}
