+++ components/item/innerPaths.ts
